conflicts: move `Merge<T>` to `merge` module
The `merge` module now seems like the obvious place for this type.
